remove authority + remove pendingtask checks from `persist_task_output`
